Removed restriction on peer data & improved safety
This fixes memory violations of storing Go pointers in C land. Instead,
this copies data into a C array and ENetPeer->data stores a C pointer to
that.

The downside here is that this then needs to be correctly freed which I
cannot find a clean way of doing as there are no underlying hooks on
peer destruction. Instead this includes documentation with a clear
warning to the user and the expectation of manual removal of stored peer
data.

Removed documentation & constant around max data length. Technically
there is still a limit & check, but it is much larger than previously.
Given the documented use case for this feature in enet is storing client
information (e.g. identifier), the new limit should be good enough.
Downside here is the 4 byte overhead, but again I don't think this will
be a big deal in practice.

Updated tests for the new implementation + made sure we run a strict cgo
check at runtime. It was this cgo check that confirmed the unsafe memory
usage previously. It is now happy.